requirements.txt
telethon==1.34.0
python-binance==1.0.19
cryptography==41.0.7
gunicorn
flask